+++ lib/features/product.dart
@@ -1,0 +1,13 @@
+class Product {
+  final String category;
+  final String name;
+  final String image;
+  final int price;
+
+  Product({
+    required this.category,
+    required this.name,
+    required this.image,
+    required this.price,
+  });
+}
