+++ lib/features/products.dart
@@ -1,78 +1,8 @@
 import 'package:flutter/material.dart';
+import 'category.dart';
+import 'product.dart';
+import 'productCard.dart';
 
-class Product {
-  final String category;
-  final String name;
-  final String image;
-  final int price;
-
-  Product({
-    required this.category,
-    required this.name,
-    required this.image,
-    required this.price,
-  });
-}
-
-class Category extends StatefulWidget {
-  final Function(int) onCategorySelected;
-  final int selectedIndex;
-
-  const Category({
-    Key? key,
-    required this.onCategorySelected,
-    required this.selectedIndex,
-  }) : super(key: key);
-
-  @override
-  State<Category> createState() => _CategoryState();
-}
-
-class _CategoryState extends State<Category> {
-  final List<String> categories = [
-    'Кофе с молоком',
-    'Черный кофе',
-    'Американо',
-    'Авторские напитки'
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 35,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemCount: categories.length,
-        itemBuilder: (context, index) {
-          return GestureDetector(
-            onTap: () {
-              widget.onCategorySelected(index);
-            },
-            child: Container(
-              padding: const EdgeInsets.only(left: 10, right: 10),
-              margin: const EdgeInsets.only(right: 10),
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(30),
-                color:
-                    widget.selectedIndex == index ? Colors.blue : Colors.white,
-              ),
-              child: Center(
-                child: Text(
-                  categories[index],
-                  style: TextStyle(
-                      fontSize: 15,
-                      color: widget.selectedIndex == index
-                          ? Colors.white
-                          : Colors.black),
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
 
 class Products extends StatefulWidget {
   const Products({super.key});
@@ -412,85 +342,6 @@ class _ProductsState extends State<Products> {
       position,
       duration: const Duration(milliseconds: 500),
       curve: Curves.easeInOut,
-    );
-  }
-}
-
-class ProductCard extends StatelessWidget {
-  final Product product;
-  final VoidCallback onAddToCart;
-  final VoidCallback onRemoveFromCart;
-  final int cartQuantity;
-  final VoidCallback onIncrease;
-  final VoidCallback onDecrease;
-
-  const ProductCard({
-    Key? key,
-    required this.product,
-    required this.onAddToCart,
-    required this.onRemoveFromCart,
-    required this.cartQuantity,
-    required this.onIncrease,
-    required this.onDecrease,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Colors.white,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Container(
-            height: 100,
-            child: Image.asset(product.image, fit: BoxFit.contain),
-          ),
-          const SizedBox(height: 15),
-          Text(
-            product.name,
-            style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 5),
-          cartQuantity > 0
-              ? Column(
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        IconButton(
-                          onPressed: onDecrease,
-                          icon: const Icon(Icons.remove),
-                        ),
-                        Text(
-                          cartQuantity.toString(),
-                          style: const TextStyle(fontSize: 20),
-                        ),
-                        IconButton(
-                          onPressed: onIncrease,
-                          icon: const Icon(Icons.add),
-                        ),
-                      ],
-                    ),
-                  ],
-                )
-              : Padding(
-                  padding: const EdgeInsets.only(left: 13, right: 13, top: 2),
-                  child: ElevatedButton(
-                    onPressed: onAddToCart,
-                    child: Text('${product.price} ₽'),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.blueAccent,
-                      foregroundColor: Colors.white,
-                      minimumSize: const Size(double.infinity, 40),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(30),
-                      ),
-                    ),
-                  ),
-                ),
-        ],
-      )
     );
   }
 }
